+++ latex.py
@@ -6,20 +6,21 @@ from database import database_exercices
 from exercises import get_exercises
 
 class latex_document():
-    def __init__(self, db: database_exercices, filename: str, numberwithin: str = "", extra_header: str = "") -> None:
+    def __init__(self, db: database_exercices, filename: str, numberwithin: str | None = "", extra_header: str = "") -> None:
         self.db = db
         self.flnm = filename
+        self.numberexs = numberwithin != None
         self.content = gen_header(numberwithin, extra_header)
     def add_toc(self) -> None:
         self.content += "\n\\tableofcontents"
-    def add_sec(self, name: str) -> None:
-        self.content += f"\n\\section{{{name}}}"
-    def add_ssec(self, name: str) -> None:
-        self.content += f"\n\\subsection{{{name}}}"
-    def add_exercises(self, exs: list[int], showdiff: bool = True, showid: bool = True, showans: bool = True, showfullinfo: bool = False, showreq: bool = False, currchap: str = "", curryear: str = "") -> None:
-        self.content += gen_exercises(self.db, map(self.db.get_exercise, exs), len(str(self.db.max_exercises())), showdiff, showid, showans, "\n\\vspace{5pt}\n%", showfullinfo, showreq, currchap, curryear)
-    def add_exercise(self, ex: int, showdiff: bool = True, showid: bool = True, showans: bool = True, showfullinfo: bool = False, showreq: bool = False, currchap: str = "", curryear: str = "") -> None:
-        self.content += gen_exercises(self.db, [self.db.get_exercise(ex)], len(str(self.db.max_exercises())), showdiff, showid, showans, "\n\\vspace{5pt}\n%", showfullinfo, showreq, currchap, curryear)
+    def add_sec(self, name: str, numbered: bool = True) -> None:
+        self.content += f"\n\\section{("" if numbered else "*")}{{{name}}}"
+    def add_ssec(self, name: str, numbered: bool = True) -> None:
+        self.content += f"\n\\subsection{("" if numbered else "*")}{{{name}}}"
+    def add_exercises(self, exs: list[int], showdiff: bool = True, showid: bool = True, showtitle: bool = True, showans: bool = True, showfullinfo: bool = False, showreq: bool = False, currchap: str = "", curryear: str = "") -> None:
+        self.content += gen_exercises(self.db, map(self.db.get_exercise, exs), len(str(self.db.max_exercises())), showdiff, showid, showtitle, showans, "\n\\vspace{5pt}\n%", showfullinfo, showreq, currchap, curryear, self.numberexs)
+    def add_exercise(self, ex: int, showdiff: bool = True, showid: bool = True, showtitle: bool = True, showans: bool = True, showfullinfo: bool = False, showreq: bool = False, currchap: str = "", curryear: str = "") -> None:
+        self.content += gen_exercises(self.db, [self.db.get_exercise(ex)], len(str(self.db.max_exercises())), showdiff, showid, showtitle, showans, "\n\\vspace{5pt}\n%", showfullinfo, showreq, currchap, curryear, self.numberexs)
     def jump_page(self) -> None:
         self.content += "\n\\null\\newpage%"
     def add_ex_table(self, title: str = "Répartition des exercices", show_toc: bool = True, fill_page: bool = True, jump_page: bool = False) -> None:
@@ -29,24 +30,15 @@ class latex_document():
         file = open(self.flnm, "w", encoding="utf-8")
         file.write(content)
         file.close()
-        
 
-def generate_latex(db: database_exercices, ex_list: Iterable[tuple[int, str, int, str, str, str, str, str]], max_exr_len: int, filename: str, showdiff: bool = True, showid: bool = True, showans: bool = True, numberwithin: str = "") -> None:
-    content = gen_header(numberwithin)
-    content += gen_exercises(db, ex_list, max_exr_len, showdiff, showid, showans, "\n\\vspace{5pt}\n%")
-    content += gen_end()
-    file = open(filename, "w", encoding="utf-8")
-    file.write(content)
-    file.close()
-
-def gen_header(numberwithin: str = "", extra_header: str = "") -> str:
+def gen_header(numberwithin: str | None = "", extra_header: str = "") -> str:
     return """
         \\documentclass[11pt]{altarticle}
         \\toggleanalysepar\\togglebornelimits
         \\usepackage{marginnote}
         \\marginparsep=10pt
         \\marginparwidth=2cm
-        \\createtheorem{exercice}{Exercice}{exo}{exercice}{""" + numberwithin + """}{true}
+        \\createtheorem{exercice}{Exercice}{exo}{exercice}{""" + (numberwithin if numberwithin else "") + """}{true}
         \\makeatletter
         \\newenvironment{corrige}[2][]%
             {%
@@ -57,7 +49,7 @@ def gen_header(numberwithin: str = "", extra_header: str = "") -> str:
                 \\begin{leftbartitle}{\\argI}%
                     \\def\\footnotetext{\\savefootnote}%
                     \\small%
-                    \\setlength{\\belowdisplayskip}{0pt}\\setlength{\\belowdisplayshortskip}{0pt}\\setlength{\\abovedisplayskip}{0pt}\\setlength{\\abovedisplayshortskip}{0pt}%
+                    \\setlength{\\belowdisplayskip}{0pt}\\setlength{\\belowdisplayshortskip}{0pt}\\setlength{\\abovedisplayskip}{0pt}\\setlength{\\abovedisplayshortskip}{0pt}\\setlength{\\dotsep}{.6ex}%
                     \\let\\textitemize\\itemize%
                     \\let\\itemize\\prvitemize%
             }{%
@@ -83,7 +75,7 @@ def diff_string(diff: int) -> str:
     else:
         return '\\Large\\filledstar' + str(diff)
 
-def gen_exercises(db: database_exercices, ex_list: Iterable[tuple[int, str, int, str, str, str, str, str]], max_exr_len: int, showdiff: bool = True, showid: bool = True, showans: bool = True, ex_sep: str = "", showfullinfo: bool = False, showreq: bool = False, currchap: str = "", curryear: str = "") -> str:
+def gen_exercises(db: database_exercices, ex_list: Iterable[tuple[int, str, int, str, str, str, str, str]], max_exr_len: int, showdiff: bool = True, showid: bool = True, showtitle: bool = True, showans: bool = True, ex_sep: str = "", showfullinfo: bool = False, showreq: bool = False, currchap: str = "", curryear: str = "", numberexs: bool = False) -> str:
     content = ""
     first = True
     for (_id, name, diff, exr, ans, year, req_chap, chap) in ex_list:
@@ -93,7 +85,7 @@ def gen_exercises(db: database_exercices, ex_list: Iterable[tuple[int, str, int,
             warn(f"Exercise {_id} has an unfinished corrected version.")
         if not first:
             content += ex_sep
-        content += f"\n\\exercice[{_id}-{currchap}-{curryear}][][][true]{{{name}}}\\marginnote{{{f'\\texttt{{[{str(_id).zfill(max_exr_len)}]}}' if showid else ''}}}{{\\reversemarginpar\\marginnote{{{diff_string(diff) if showdiff else ''}}}}}\\emph{{{(', '.join(map(lambda x: db.get_chapter(int(x)).lower(), (filter(lambda x: x != '' and x != currchap, req_chap.split(',')))))) if showreq else ""}}}\\par\\nobreak%"
+        content += f"\n\\exercice{("" if numberexs else "*")}[{_id}-{currchap}-{curryear}][][][true]{{{(name if showtitle else "")}}}\\marginnote{{{f'\\texttt{{[{str(_id).zfill(max_exr_len)}]}}' if showid else ''}}}{{\\reversemarginpar\\marginnote{{{diff_string(diff) if showdiff else ''}}}}}\\emph{{{(', '.join(map(lambda x: db.get_chapter(int(x)).lower(), (filter(lambda x: x != '' and x != currchap, req_chap.split(',')))))) if showreq else ""}}}\\par\\nobreak%"
         req_chap_str = ', '.join(map(lambda x: db.get_chapter(int(x)), (filter(lambda x, chap=currchap: x != '' and x != chap, req_chap.split(',')))))
         if showfullinfo:
             content += f"\n\\begingroup\\slshape Année(s): {', '.join(map(lambda x: db.get_year(int(x)), (filter(lambda x: x != '', year.split(',')))))}\\\\Chapitre(s): {', '.join(map(lambda x: db.get_chapter(int(x)), (filter(lambda x: x != '', chap.split(',')))))} {('\\\\Chapitre(s) nécessaire(s): ' + req_chap_str) if req_chap_str else ''}\\endgroup%\n"
@@ -161,6 +153,7 @@ def generate_exercise_sheet(ex_list_1: Iterable[tuple[int, str, int, str, str, s
         \\ExplSyntaxOff
         \\makeatother
         \\hideitems
+        \\newcommand{\\indic}[1]{\\vspace{-\\baselineskip}}
     """
     content += f"""
         \\begin{{document}}{{\\color{{lightgray}}\\hrule}}\\vfill{'\\hrule' if showminipagemars else ''}\\stepcounter{{section}}\\begin{{minipage}}[t][8.5cm][t]{{\\textwidth}}%
@@ -193,8 +186,8 @@ def generate_exercise_sheet(ex_list_1: Iterable[tuple[int, str, int, str, str, s
     file.write(content)
     file.close()
 
-def gen_book(db: database_exercices, flnm: str, numberwithin: str = "subsection", showans: bool = False, showdiff: bool = True) -> None:
-    doc = latex_document(db, flnm, numberwithin)
+def gen_book(db: database_exercices, flnm: str, numberwithin: str | None = "subsection", showans: bool = False, showdiff: bool = True, showid: bool = True, showtitle: bool = True, showreq: bool = False) -> None:
+    doc = latex_document(db, flnm, numberwithin, "\\newcommand{\\indic}[1]{\\textsl{#1}}\n")
     doc.add_toc()
     doc.add_ex_table(fill_page=False, jump_page=True)
     for y_id, yname in [("", "Non répertorié"),("2","MPS/2I"),("1","MP(I)")]: # À modifier, garder l'ordre
@@ -209,10 +202,10 @@ def gen_book(db: database_exercices, flnm: str, numberwithin: str = "subsection"
                 if not c:
                     doc.add_ssec(cname)
                     c = True
-                doc.add_exercise(x, showdiff=showdiff, showans=showans, showreq=True, currchap=str(c_id), curryear=str(y_id))
+                doc.add_exercise(x, showdiff=showdiff, showans=showans, showid=showid, showreq=showreq, showtitle=showtitle, currchap=str(c_id), curryear=str(y_id))
     doc.gen_doc()
 
-def gen_exercise_book(db: database_exercices, flnm: str, ids: str, showans: bool = False, showdiff: bool = False) -> None:
+def gen_exercise_book(db: database_exercices, flnm: str, ids: str, showans: bool = False, showdiff: bool = False, showid: bool = True, showtitle: bool = True, showreq: bool = False, numberexs: bool = True) -> None:
     extra_hd = """
         \\usepackage{xparse}
         \\makeatletter
@@ -247,9 +240,61 @@ def gen_exercise_book(db: database_exercices, flnm: str, ids: str, showans: bool
         \\ExplSyntaxOff
         \\makeatother
         \\hideitems
+        \\newcommand{\\indic}[1]{\\textsl{#1}}
     """
     doc = latex_document(db, flnm, extra_header=extra_hd)
-    doc.content += gen_exercises(db, get_exercises(doc.db, ids), len(str(doc.db.max_exercises())), showans=showans, showdiff=showdiff)
+    doc.content += gen_exercises(db, get_exercises(doc.db, ids), len(str(doc.db.max_exercises())), showans=showans, showdiff=showdiff, showid=showid, showtitle=showtitle, showreq=showreq, numberexs=numberexs)
+    doc.gen_doc()
+
+def gen_recap(db: database_exercices, flnm: str, ids: tuple[list[str], str], showans: bool = False, showdiff: bool = False, showid: bool = True, showtitle: bool = True, showreq: bool = False, numberexs: bool = True) -> None:
+    extra_hd = """
+        \\usepackage{xparse}
+        \\makeatletter
+        \\ExplSyntaxOn
+        \\NewDocumentCommand{\\hideit}{}
+        {
+        \\gaweiliex_hide:
+        }
+        \\NewDocumentCommand{\\hideitems}{}
+        {
+        \\bool_set_true:N \\l_gaweiliex_hide_bool
+        }
+        \\NewDocumentCommand{\\showitems}{}
+        {
+        \\bool_set_false:N \\l_gaweiliex_hide_bool
+        }
+        \\bool_new:N \\l_gaweiliex_hide_bool
+        \\cs_new_protected:Nn \\gaweiliex_hide:
+        {
+        \\bool_if:NT \\l_gaweiliex_hide_bool
+        {
+            \\stepcounter{\\@enumctr}%
+            \\item[]\\vspace{-\\baselineskip}%
+            \\peek_regex_replace_once:nn
+            % search \\item followed by anything until finding
+            % \\item or \\hideit or \\end{<current environment>}
+            { \\c{item}.*?(\\c{item}|\\c{hideit}|\\c{end}\\{\\u{@currenvir}\\}) }
+            % replace by the matching item
+            { \\1 }
+        }
+        }
+        \\ExplSyntaxOff
+        \\makeatother
+        \\hideitems
+        \\title{Colle \\no{} ? -- Classe}
+        \\author{}
+        \\date{DD/MM/YYYY}
+        \\renewcommand{\\theexercice}{\\arabic{exercice}}
+        \\newcommand{\\indic}[1]{\\vspace{-\\baselineskip}}
+    """
+    doc = latex_document(db, flnm, extra_header=extra_hd, numberwithin="section")
+    doc.content += "\\maketitle\n"
+    for student in range(3):
+        doc.add_sec(f"\\stepcounter{{section}}Élève \\textsc{{{student + 1}}} (??/20)", False)
+        doc.content += gen_exercises(db, get_exercises(doc.db, ids[0][student]), len(str(doc.db.max_exercises())), showans=showans, showdiff=showdiff, showid=showid, showtitle=showtitle, showreq=showreq, numberexs=numberexs)
+        doc.content += "\\null\n\\dots\n\n"
+    doc.add_sec(f"\\stepcounter{{section}}Extra", False)
+    doc.content += gen_exercises(db, get_exercises(doc.db, ids[1]), len(str(doc.db.max_exercises())), showans=showans, showdiff=showdiff, showid=showid, showtitle=showtitle, showreq=showreq, numberexs=numberexs)
     doc.gen_doc()
 
 def generate_exercises_table(db: database_exercices, title: str = "Répartition des exercices", show_toc: bool = True, fill_page: bool = True, jump_page: bool = False) -> str: